requirements.txt
playwright>=1.53.0
requests>=2.32.0
Markdown>=3.7
pdfkit>=1.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
